requirements.txt
# ==deps.py #002
fastapi
uvicorn
passlib[bcrypt]
pyjwt
python-multipart
